+++ pandapipes/__init__.py
@@ -15,6 +15,6 @@ from pandapipes.create import *
 from pandapipes.io.file_io import *
 from pandapipes.pipeflow import *
 from pandapipes.toolbox import *
-from pandapipes.pipeflow_setup import *
+from pandapipes.pf.pipeflow_setup import *
 from pandapipes.std_types import *
 import pandapipes.plotting
