+++ pandapipes/component_models/abstract_models/base_component.py
@@ -2,7 +2,7 @@
 # and Energy System Technology (IEE), Kassel, and University of Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-from pandapipes.component_models.auxiliaries.component_toolbox import init_results_element
+from pandapipes.component_models.component_toolbox import init_results_element
 
 try:
     import pplog as logging
