+++ pandapipes/component_models/circulation_pump_pressure_component.py
@@ -6,8 +6,8 @@ import numpy as np
 from numpy import dtype
 from pandapipes.component_models.abstract_models import CirculationPump
 from pandapipes.idx_node import PINIT, NODE_TYPE, P, EXT_GRID_OCCURENCE
-from pandapipes.internals_toolbox import _sum_by_group
-from pandapipes.pipeflow_setup import get_lookup
+from pandapipes.pf.internals_toolbox import _sum_by_group
+from pandapipes.pf.pipeflow_setup import get_lookup
 
 try:
     import pplog as logging
