+++ pandapipes/component_models/heat_exchanger_component.py
@@ -7,7 +7,7 @@ from numpy import dtype
 from pandapipes.component_models.abstract_models import BranchWZeroLengthComponent
 from pandapipes.idx_branch import PL, TL, ALPHA, \
     TEXT, QEXT, T_OUT, D, AREA, LOSS_COEFFICIENT as LC
-from pandapipes.pipeflow_setup import get_fluid
+from pandapipes.pf.pipeflow_setup import get_fluid
 
 try:
     import pplog as logging
