+++ pandapipes/component_models/junction_component.py
@@ -7,10 +7,10 @@ from warnings import warn
 import numpy as np
 from numpy import dtype
 from pandapipes.component_models.abstract_models import NodeComponent
-from pandapipes.component_models.auxiliaries.component_toolbox import p_correction_height_air
+from pandapipes.component_models.component_toolbox import p_correction_height_air
 from pandapipes.idx_node import L, ELEMENT_IDX, RHO, PINIT, node_cols, HEIGHT, TINIT, PAMB, \
     ACTIVE as ACTIVE_ND
-from pandapipes.pipeflow_setup import add_table_lookup, get_table_number, \
+from pandapipes.pf.pipeflow_setup import add_table_lookup, get_table_number, \
     get_lookup
 from pandapipes.properties.fluids import get_fluid
