+++ pandapipes/pf/__init__.py
@@ -2,6 +2,8 @@
 # and Energy System Technology (IEE), Kassel, and University of Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-from .derivative_toolbox import *
-from .build_system_matrix import *
-from .component_toolbox import *
+
+from pandapipes.pf.pipeflow_setup import *
+from pandapipes.pf.internals_toolbox import *
+from pandapipes.pf.pipeflow_setup import *
+from pandapipes.pf.build_system_matrix import *
