+++ pandapipes/test/api/test_aux_function.py
@@ -2,17 +2,8 @@
 # and Energy System Technology (IEE), Kassel, and University of Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-import os
-
 import numpy as np
-import pandapipes
-import pandas as pd
-import pytest
-from pandapipes.component_models import Pipe, Junction
-from pandapipes.idx_node import PINIT, TINIT
-from pandapipes.pipeflow_setup import get_lookup
-from pandapipes.test.pipeflow_internals import internals_data_path
-from pandapipes.internals_toolbox import select_from_pit
+from pandapipes.pf.internals_toolbox import select_from_pit
 
 
 def test_select_from_pit():
